+++ test/utilities.py
@@ -1,2 +1,4 @@
+import numpy as np
+
 def assert_equal(value, target, precision=1e-3):
-    assert value > target-precision and value < target+precision
+    assert np.all(value > target-precision) and np.all(value < target+precision)