+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * includes `clip_low` and `clip_high` arguments in viewing methods
 * Add `save_image` to allow saving a single image
 * Add `segment_module_part` as a more flexible variant of `segment_module` and `segment_cell`
+* Add `from_other` method to `Image` and `ImageSequence` to allow partial overwriting of attributes
 
 ## 0.0.1
 * Initial version